Update nextflow_schema.json: remove fasta from required (#218)
This should fix a small bug when running the pipeline via Seqera Platform. Because the parameter is listed as required, pipelines cannot be submitted unless it is set. Even though this shouldn't be required since `--genome` is set. I am guessing nf-schema itself might have similar issues, but haven't tested it